requirements.txt
dj-database-url==0.4.2
gunicorn==19.7.1
whitenoise==3.3.0
django==1.10.5
djangorestframework==3.6.2
django-rest-swagger
